Move betty.plugins.document_provider to betty.document_providers
